change the way of windows service creation.
Powershell script shipped with package expects configuration file and
executable in one folder. This does not allow scenarios when
configuration folder is overrided to be in a parent folder, for
example. Proposed block creates service the same way as default script
but takes configuration file and executable directly from attributes.